refactor(backend): extract edit mode into providers/generate_modes/edit.py
Move the inline `if body.mode == "edit"` branch out of generate._event_stream
into `stream_edit(...)`, an async generator that yields byte-identical SSE
frames. Mirrors the merged ascend extraction (#113): cross-module providers
imported at module top; generate.py's own helpers (`_sse`,
`_abort_if_disconnected`, `_condition_url_for_role`) threaded in as kwargs; no
reach-back into module globals.

generate.py keeps a thin dispatch that forwards frames and returns. Same flag
gates (EDIT_REGION / EDIT_JUDGE), same early-returns, same spend accounting,
same error frames — the moved body is a diff-verified byte-for-byte copy. Drop
the now-unused `EditAttempt` TYPE_CHECKING import from generate.py.
