+++ apps/modal-backend/providers/generate_modes/__init__.py
@@ -4,3 +4,5 @@ Each handler is an async generator that yields the exact same `_sse(...)` frames
 the inline branch used to, threading generate.py's stream helpers explicitly so
 this package never reaches back into generate.py's module globals.
 """
+
+from providers.generate_modes.edit import stream_edit as stream_edit
